agent/requirements.txt
bedrock-agentcore==1.4.8
strands-agents
